pyproject.toml
[tool.poetry]
name = "grimoire-ssg"
packages = [
    { include = "grimoiressg" }
]
version = "0.1.0"
description = "A minimalistic Static Site Generator"
authors = ["Sigma <git@sigma-star.io>"]
license = "BSD-2-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
markdown = "^3.7"
jinja2 = "^3.1.5"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
